Update location_list to allow stopover argument
Switching the arg input from a list to dictionary with the format provided by Google for passing way points. This allows setting a 'stopover' value in the dictionary which sets the way point to either a stop over point or a via point. Note: I've only updated the function with this commit and no other documentation outside of the documentation inside the function definition.