Set base_url default value inside _request() instead of in signature
If an end-user of the library wants to override _DEFAULT_BASE_URL (e.g. in order to use a proxy), this makes it easier. The user can simply run

```
googlemaps.client._DEFAULT_BASE_URL = 'http://my_proxy_dns'
```

in their server/script's startup code.

Without this change, the default value for the parameter gets "baked" into the function definition and is harder to change.