Fix the documentation for find_place.
In the documentation for find_place, it says that `fields` parameter is "separated by a comma",
but it is in fact a list, instead of a comma separated string. Perhaps I just misunderstood
what it meant? I think removing the "separated by a comma" would be clearer.

The find_place documentation also says that the type for `input` is list, and
there's no type specified for `fields`. In reality (upon reading the code),
I noticed that the `input` is a string, and `fields` is a list type.

I've updated the documentation to reflect correct types.